Return from GPSDriverUBX::receive when ready, don't wait until no data is received
The receive implementation was waiting until no data was read for a timeout period before checking
if it is ready to return. However, when the bus utilization is high, there might be no gap in the
received data stream and as a result, the receive timeout is triggered. The driver is then
reinitialized and potentially doesn't start again since the GPS device is still sending data at high
rates and the configure logic times out due to the same check. The timeout was recently increased
which makes it more likely that this scenario happens.
The receive function returns now immediately when both posllh and velned are received and doesn't
wait until there is no data read.