ubx: add function for disabling/re-enabling mb rover operation
A receiver in moving base rover mode will output data with delay up to 1 second if it cannot compute a relative solution. This can for example happen if the moving base experiences degraded signal or power loss.

In such a scenario, the rover might be the only receiver left, and for redundancy it's important to bring it back into a normal state without delay in order to avoid toilet bowl/position loss.

Testing with an F9P receiver, the quickest way I found to achieve this is to send a soft position reset.

NB! This is only meant for serial GPSes connected to one another over
UART2